updated 6/25/2020

+++ Unknown/countdown.cpp
@@ -1,7 +1,7 @@
 #include<bits/stdc++.h>
 using namespace std;
 
-using ll = long long;
+using ll = int int;
 
 int  main(){
     int t;
